added kecleon shop!

+++ src/item_gen.c
@@ -69,7 +69,7 @@ static void ShuffleArrayU16(u16* array, u32 size)
     // Code from https://stackoverflow.com/questions/6127503/shuffle-array-in-c.
     for (i = 0; i < size - 1; ++i) 
     {
-        j = i + RandomF() / (UINT16_MAX / (size - i) + 1);
+        j = i + RandomF() / (UINT32_MAX / (size - i) + 1);
         t = array[j];
         array[j] = array[i];
         array[i] = t;
